extracted ISigner interface from registries/types.ts
to fix circular dependancies
