+++ ts/identity/didDocument/didDocument.ts
@@ -3,7 +3,7 @@ import { IDidDocumentAttrs } from './types'
 import { canonize } from 'jsonld'
 import { EcdsaLinkedDataSignature } from '../../linkedDataSignature'
 import { AuthenticationSection, PublicKeySection, ServiceEndpointsSection } from './sections'
-import { ISigner } from '../../registries/types'
+import { ISigner } from '../../registries/signerType'
 import { ContextEntry } from 'cred-types-jolocom-core'
 import { defaultContextIdentity } from '../../utils/contexts'
 import { sha256, publicKeyToDID } from '../../utils/crypto'
