+++ ts/registries/signerType.ts
@@ -1,0 +1,4 @@
+export interface ISigner {
+    did: string
+    keyId: string
+}
